Rename TicTacToe files for clarity and refactor bot logic
- Updated imports and paths to reflect new file structure
- Improved bot logic with refactoring; all tests passing
