+++ src/tkinteros/applications/tictactoe/bot.py
@@ -2,6 +2,8 @@ import time
 import logging
 import random
 
+import tkinteros.applications.tictactoe.game_logic as game_logic
+
 
 class TicTacToeBot:
     def __init__(self, ai_symbol: str = "O", player_symbol: str = "X", random_start: bool = False,
@@ -16,91 +18,19 @@ class TicTacToeBot:
 
     def move(self, board: list[list[str]]):
         self.times_ran = 0
-        if self.random_start and self.get_move_count(board) < 3:
+        if self.random_start and game_logic.get_move_count(board) < 3:
             while True:
                 random_move = (random.randint(0, len(board) - 1), random.randint(0, len(board) - 1))
                 if board[random_move[0]][random_move[1]] == " ":
                     return random_move
 
         return self.best_move(board)
-    
-
-    def get_move_count(self, board: list[list[str]]) -> int:
-        count = 0
-        for row in board:
-            for cell in row:
-                count += 1 if cell != self.empty_cell else 0
-        return count
-    
-
-    def evaluate_winner(self, window: list[str]):
-        window_set = set(window)
-        if len(window_set) == 1 and list(window_set)[0] != self.empty_cell:
-            return list(window_set)[0]
-        return None
-    
-
-    def get_board_windows(self, board: list[list[str]]) -> list[list[str]]:
-        board_size = len(board)
-        window_size = self.points_to_win
-        windows = []
-
-        # Horizontal
-        for row in range(board_size):
-            for col in range(board_size - window_size + 1):
-                windows.append(board[row][col:col+window_size])
-
-        # Vertical
-        for col in range(board_size):
-            for row in range(board_size - window_size + 1):
-                window = []
-                for k in range(window_size):
-                    window.append(board[row + k][col])
-                windows.append(window)
-
-        # Diagonal (\)
-        for row in range(board_size - window_size + 1):
-            for col in range(board_size - window_size + 1):
-                window = []
-                for k in range(window_size):
-                    window.append(board[row + k][col + k])
-                windows.append(window)
-
-        # Diagonal (/)
-        for row in range(window_size - 1, board_size):
-            for col in range(board_size - window_size + 1):
-                window = []
-                for k in range(window_size):
-                    window.append(board[row - k][col + k])
-                windows.append(window)
-
-        return windows
-
-
-    def evaluate_game_state(self, board: list[list[str]]) -> str:
-        """
-            Checks if the game is over and returns the symbol of the play that won.
-            Returns None if the game is not over and returns "tie" if all cells are filled.
-        """
-        windows = self.get_board_windows(board)
-
-        for window in windows:
-            winner = self.evaluate_winner(window)
-            if winner:
-                return winner
-            
-        # Check if the board is full
-        is_full = all(cell != self.empty_cell for row in board for cell in row)
-        if is_full:
-            return "tie"
-
-        return None  # Game still in progress
 
 
     def heuristic(self, board: list[list[str]]):
         score = 0
 
-        windows = self.get_board_windows(board=board)
+        windows = game_logic.get_board_windows(board=board)
 
         for window in windows:
             if window.count(self.ai_symbol) == self.points_to_win - 1 and window.count(self.empty_cell) == 1:
@@ -124,7 +54,7 @@ class TicTacToeBot:
     def minimax(self, board, is_maximizing: bool, depth: int = 0, 
                 alpha: float = -float("inf"), beta: float = float("inf")):
         self.times_ran += 1
-        result = self.evaluate_game_state(board)
+        result = game_logic.evaluate_game_state(board)
         if result == self.ai_symbol:
             return 1 / (depth + 1)  # prefer quicker wins
         elif result == self.player_symbol:
